Crate libsignal_core

Modules

curve

Structs

DeviceId

The type used in memory to represent a device, i.e. a particular Signal client instance which represents some user.

E164

InvalidDeviceId

Error for trying to construct a DeviceId with an invalid value.

ProtocolAddress

Represents a unique Signal client instance as (<user ID>, <device ID>) pair.

WrongKindOfServiceIdError

The error returned for a failed “downcast” conversion from a ServiceId to a specific kind of service ID (e.g. Pni).

Enums

ServiceId

A Signal service ID, which can be one of various types.

ServiceIdKind

Known types of ServiceId.

Constants

VERSION

Functions

try_scoped

Simple wrapper that invokes a lambda.

Type Aliases

Aci

A service ID representing an ACI (“ACcount Identifier”).

Pni

A service ID representing a PNI (“Phone Number Identifier”).

ServiceIdFixedWidthBinaryBytes

The fixed-width binary representation of a ServiceId.